refact: ArchiveMonth, ArchiveDay를 ArchiveDate로 통합
